move plugins to new structure
