+++ pype/hosts/celaction/plugins/publish/collect_render_path.py
